+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+- Add support for the XLIFF localization interchange format:
+  https://docs.oasis-open.org/xliff/v1.2/os/xliff-core.html
+
+- [BREAKING] Replaced `xlbDir` config file property with `interchange` property.
+  The interchange format is set with `interchange.format` (currently `xliff` or
+  `xlb`), and other format-specific configuration is set in that object.
 
 ## [0.1.2] - 2020-05-09